Move functions realated to getting/reading files into file_acquisition.py
